HaikuDepot : Fix Compile Warnings
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7298
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/apps/haikudepot/server/WebAppInterface.cpp
@@ -68,13 +68,11 @@ public:
 	{
 	}
 
-	virtual	void DownloadProgress(BUrlRequest* caller, off_t bytesReceived,
-		ssize_t bytesTotal)
+	virtual	void DownloadProgress(BUrlRequest* caller, off_t bytesReceived, off_t bytesTotal)
 	{
 	}
 
-	virtual void UploadProgress(BUrlRequest* caller, off_t bytesSent,
-		ssize_t bytesTotal)
+	virtual void UploadProgress(BUrlRequest* caller, off_t bytesSent, off_t bytesTotal)
 	{
 	}
